FDP-387: fixed Sonar issues 'Cognitive Complexity of methods should not be too high' and several other issues in affected classes. InIec61850...Command classes, refactored the used Function<Void> from within one single method into a new class. Renewed suppression of complexity issue in OslpChannelHandler in web-device-simulator after Sonar issue name seems to have changed.
